Add : Next.js 버전업 && 몇몇 css 변경
