+++ nextapp/package.json
@@ -18,11 +18,12 @@
     "eslint": "8.41.0",
     "eslint-config-next": "latest",
     "html-react-parser": "^4.2.2",
-    "next": "latest",
+    "next": "canary",
     "prismjs": "^1.29.0",
     "react": "latest",
     "react-dom": "latest",
     "react-icons": "^4.10.1",
+    "react-lottie-player": "^1.5.5",
     "styled-components": "^6.0.5",
     "supports-color": "^9.4.0",
     "swr": "^2.2.2",
