+++ nextapp/src/components/modules/Post/EditPost.tsx
@@ -59,6 +59,26 @@ const Label = styled.label`
 	margin-right: 10px;
 `;
 
+const TextInput = styled.input`
+	border-radius: 5px; 
+	padding:5px;
+	border:1px solid gray; 
+	font-size: 0.7rem; 
+	color: '#5e5e5e';
+	&::placeholder{
+		color: #b8b8b8;
+	}
+`;
+
+const Select = styled.select`
+	border-radius: 5px; 
+	border:1px solid gray; 
+	font-size: 0.7rem; 
+	height: 30px; 
+	font-family: inherit;
+	color: #5e5e5e;
+`;
+
 export default function EditPost ({ children }: {children: React.ReactNode}) {
 	const json = window.sessionStorage.getItem('post-storage');
 	const data: PostType = JSON.parse(json as string).state.nowPost;
@@ -133,8 +153,7 @@ export default function EditPost ({ children }: {children: React.ReactNode}) {
 			<EditorBox>
 				<InfoBox>
 					<Label>글 제목</Label>
-					<input 
-						style={{borderRadius: '5px', padding:'5px', border:'1px solid gray', fontSize:'0.7rem', opacity:'0.5'}} 
+					<TextInput 
 						type="text" 
 						defaultValue={data.title}
 						onChange={(e)=>{setTitle(e.target.value)}}
@@ -142,12 +161,11 @@ export default function EditPost ({ children }: {children: React.ReactNode}) {
 				</InfoBox>
 				<InfoBox>
 					<Label>카테고리 선택하기</Label>
-					<select 
-						style={{borderRadius: '5px', border:'1px solid gray', height:'30px', fontFamily:'inherit',color: '#5e5e5e',opacity:'0.5'}}
+					<Select 
 						onChange={(e)=>{setCategoryId(e.target.value)}}
 					>
 						{optionList}
-					</select>
+					</Select>
 				</InfoBox>
 				<Editor
 					ref={editorRef}
