--- nextapp/src/hooks/GoogleAnalytics.tsx
@@ -1,32 +1,0 @@
-'use client';
-
-import Script from 'next/script'
-
-export default function GoogleAnalytics({GA_TRACKING_ID} : {GA_TRACKING_ID : string}){
-    return (
-        <>
-            <Script 
-							strategy="afterInteractive" 
-							src={`https://www.googletagmanager.com/gtag/js?id=${GA_TRACKING_ID}`}
-						/>
-            <Script 
-							id='google-analytics' 
-							strategy="afterInteractive"
-							dangerouslySetInnerHTML={{
-                __html: `
-                window.dataLayer = window.dataLayer || [];
-                function gtag(){dataLayer.push(arguments);}
-                gtag('js', new Date());
-
-                gtag('consent', 'default', {
-                    'analytics_storage': 'denied'
-                });
-                
-                gtag('config', '${GA_TRACKING_ID}', {
-                    page_path: window.location.pathname,
-                });
-                `,
-							}}
-            />
-        </>
-)}
